Add actions package and move messages there
