+++ AndroidSDKCore/src/main/java/com/leanplum/messagetemplates/actions/ConfirmMessage.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014, Leanplum, Inc. All rights reserved.
+ * Copyright 2020, Leanplum, Inc. All rights reserved.
  *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.leanplum.messagetemplates.controllers;
+package com.leanplum.messagetemplates.actions;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -33,33 +33,35 @@ import com.leanplum.messagetemplates.MessageTemplateConstants.Args;
 import com.leanplum.messagetemplates.MessageTemplateConstants.Values;
 
 /**
- * Registers a Leanplum action that displays a system alert dialog.
+ * Registers a Leanplum action that displays a system confirm dialog.
  *
  * @author Andrew First
  */
-public class AlertMessage {
+public class ConfirmMessage {
 
   public static ActionArgs createActionArgs(Context context) {
     return new ActionArgs()
         .with(Args.TITLE, Util.getApplicationName(context))
-        .with(Args.MESSAGE, Values.ALERT_MESSAGE)
-        .with(Args.DISMISS_TEXT, Values.OK_TEXT)
-        .withAction(Args.DISMISS_ACTION, null);
+        .with(Args.MESSAGE, Values.CONFIRM_MESSAGE)
+        .with(Args.ACCEPT_TEXT, Values.YES_TEXT)
+        .with(Args.CANCEL_TEXT, Values.NO_TEXT)
+        .withAction(Args.ACCEPT_ACTION, null)
+        .withAction(Args.CANCEL_ACTION, null);
   }
 
   public static void showMessage(ActionContext context) {
     Activity activity = LeanplumActivityHelper.getCurrentActivity();
-    if (activity == null || activity.isFinishing()) {
+    if (activity == null || activity.isFinishing())
       return;
-    }
 
     new AlertDialog.Builder(activity)
         .setTitle(context.stringNamed(Args.TITLE))
         .setMessage(context.stringNamed(Args.MESSAGE))
         .setCancelable(false)
-        .setPositiveButton(
-            context.stringNamed(Args.DISMISS_TEXT),
-            (dialog, id) -> context.runActionNamed(Args.DISMISS_ACTION))
+        .setPositiveButton(context.stringNamed(Args.ACCEPT_TEXT),
+            (dialog, id) -> context.runTrackedActionNamed(Args.ACCEPT_ACTION))
+        .setNegativeButton(context.stringNamed(Args.CANCEL_TEXT),
+            (dialog, id) -> context.runActionNamed(Args.CANCEL_ACTION))
         .create()
         .show();
   }
